Removing localhost check from config.php
This prevents config.php being accessible from outside localhost to protect against the scenario where the file is accidentally deployed to production.  The Dockerfile should take care of this, and the configuration build into the NGINX image should also prevent access to this file, so we shouldn't need this protection.  Having this check in place prevents us from using config.php from the host machine, because requests from it don't come from localhost.