Updating to match version installed by composer
Since performing the original install of Symfony there has been an update to the symfony_requirements file - as this file has been committed in Git, this means performing a `composer install` in this repository results in a diff showing on this file.  Potentially the file should just be treated like the vendor directory, ignored from the Git repo, seeing as it gets re-instated by composer install anyway.  But the .gitignore specified in app/www/vendor/sensio/distribution-bundle/Composer/ScriptHandler.php specifically ensures that symfony_requirements is not ignored, so it seems the intention is for this file to be kept in Git.  So for now I'm just updating the source-controlled file so the changes don't show up, but in the future maybe it should just be removed.